Split book_rental file into 2
