+++ book.rb
@@ -11,16 +11,3 @@ class Book
     Rental.new(date, self, person)
   end
 end
-
-class Rental
-  attr_accessor :date
-  attr_reader :book, :person
-
-  def initialize(date, book, person)
-    @date = date
-    @book = book
-    book.rentals << self
-    @person = person
-    person.rentals << self
-  end
-end
