+++ student.rb
@@ -1,5 +1,6 @@
 require_relative 'person'
-require_relative 'book_rental'
+require_relative 'book'
+require_relative 'rental'
 require_relative 'classroom'
 
 class Student < Person
